Fixed bug that prevented extraction of files with 0 filesize. Fixed bug that kept local files open after importing them and saving image.

+++ YaffsItem.h
@@ -111,7 +111,7 @@ public:
     QString getAlias() const { return mYaffsObjectHeader.alias; }
     int getHeaderPosition() const { return mHeaderPosition; }
     const yaffs_obj_hdr& getHeader() const { return mYaffsObjectHeader; }
-    int getFileSize() const { return mYaffsObjectHeader.file_size_low; }
+    size_t getFileSize() const { return mYaffsObjectHeader.file_size_low; }
     uint getUserId() const { return mYaffsObjectHeader.yst_uid; }
     uint getGroupId() const { return mYaffsObjectHeader.yst_gid; }
     uint getPermissions() const { return mYaffsObjectHeader.yst_mode; }
